mv figshare.py to remote/figshare.py and extract download_file from data.py to new remote/fetch.py
