+++ scripts/upload_data_files_to_figshare.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from tqdm import tqdm
 
-import matbench_discovery.figshare as figshare
+import matbench_discovery.remote.figshare as figshare
 from matbench_discovery import DATA_DIR, PKG_DIR, ROOT
 from matbench_discovery.data import round_trip_yaml
 from matbench_discovery.enums import DataFiles
